+++ memapi/examples/memory.rs
@@ -1,0 +1,12 @@
+/*use pymemprofile_api::oom::{MemoryInfo, RealMemoryInfo};
+
+fn main() {
+    let rm = RealMemoryInfo::new();
+    println!(
+        "Available memory: {}, resident process memory: {}",
+        rm.get_available_memory(),
+        rm.get_resident_process_memory()
+    );
+}*/
+
+fn main() {}